Refactored Player Status to contain all data evaluation involving sanity/mental state
Narrative manager now has a serializefield with player status type to be able to manage a player status
