+++ Assets/ICT371 Project/Scripts/activities_and_days/NarrativeSystem.cs
@@ -1,7 +1,5 @@
-using System;
 using System.Collections;
 using System.Collections.Generic;
-using JetBrains.Annotations;
 using UnityEngine;
 using UnityEngine.Events;
 using UnityEngine.iOS;
@@ -14,27 +12,14 @@ public class NarrativeSystem : MonoBehaviour
     UnityEvent _onDayEnd;
     [SerializeField] BookInterface _bookInterface;
     [SerializeField] List<Day> _days;
-    
-    public enum CognitiveStage
-    {
-        Early,
-        Middle,
-        Late,
-        Deceased
-    }
+
+    [SerializeField] PlayerStatus _playerInterface;
 
     // Narrative properties
     // -----------------------------------------------------------------
     int _currentDayIndex = 0;
     // -----------------------------------------------------------------
 
-    // Cognitive decay modifiers
-    // -----------------------------------------------------------------
-    private const double k_MinDecay = 0.0d, k_MaxDecay = 1.0d;
-    private double _decay = 0.0d;
-    private double _decayRate = -0.8d;
-    private CognitiveStage _stage = CognitiveStage.Early;
-    // -----------------------------------------------------------------
 
     void Awake()
     {
@@ -98,47 +83,9 @@ public class NarrativeSystem : MonoBehaviour
 
         _days[_currentDayIndex++].EndDay();
         
-        EvaluateDecay();
-        EvaluateStage();
+        _playerInterface.EvaluateDecay(_currentDayIndex);
+        _playerInterface.EvaluateStage();
 
         _onDayEnd.Invoke();
-    }
-
-    public double GetCognitiveDecay()
-    {
-        return _decay;
-    }
-
-    public CognitiveStage GetCognitiveStage()
-    {
-        return _stage;
-    }
-
-    private void EvaluateDecay()
-    {
-        // TODO: Add modifiers into decay rate
-        // Evaluate the (daily) cognitive decay s.t. f(x) = c exp(kx)
-        _decay = k_MaxDecay * Math.Exp(_decayRate * _currentDayIndex);
-        Debug.Log("Decay: " + _decay + ", Day: " + _currentDayIndex);
-    }
-
-    private void EvaluateStage()
-    {
-        if (_decay > 0.3d)
-        {
-            _stage = CognitiveStage.Middle;
-        }
-
-        if (_decay > 0.6d)
-        {
-            _stage = CognitiveStage.Late;
-        }
-
-        if (_decay > k_MaxDecay)
-        {
-            _stage = CognitiveStage.Deceased;
-        }
-
-        Debug.Log("Stage: " + _stage);
     }
 }
