+++ Assets/ICT371 Project/Scripts/player_status.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 624a2d1602e593e4494489b7b55e98f8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
