+++ Assets/ICT371 Project/Scripts/player_status/PlayerStatus.cs
@@ -1,33 +1,58 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+public enum CognitiveStage
+{
+    Early,
+    Middle,
+    Late,
+    Deceased
+};
+
 public class PlayerStatus : MonoBehaviour
 {
-    private float sanity;
+    private const double k_MinDecay = 0.0d, k_MaxDecay = 1.0d;
+    private double _decay = 0.0d;
+    private double _decayRate = -0.8d;
+    private CognitiveStage _stage = CognitiveStage.Early;
 
-    PlayerStatus(float initSanity)
+     public double GetCognitiveDecay()
     {
-        sanity = initSanity;
+        return _decay;
     }
 
-    public void reduceSanity(float decrease)
+    public CognitiveStage GetCognitiveStage()
     {
-        sanity -= decrease;
+        return _stage;
     }
 
-    public void increaseSanity(float increase)
+    public void EvaluateDecay(int currentDayIndex)
     {
-        sanity += increase;
+        // TODO: Add modifiers into decay rate
+        // Evaluate the (daily) cognitive decay s.t. f(x) = c exp(kx)
+        _decay = k_MaxDecay * Math.Exp(_decayRate * currentDayIndex);
+        Debug.Log("Decay: " + _decay + ", Day: " + currentDayIndex);
     }
 
-    public void setSanity(float sanityValue)
+    public void EvaluateStage()
     {
-        sanity = sanityValue;
-    }
+        if (_decay > 0.3d)
+        {
+            _stage = CognitiveStage.Middle;
+        }
 
-    public float getSanity()
-    {
-        return sanity;
+        if (_decay > 0.6d)
+        {
+            _stage = CognitiveStage.Late;
+        }
+
+        if (_decay > k_MaxDecay)
+        {
+            _stage = CognitiveStage.Deceased;
+        }
+
+        Debug.Log("Stage: " + _stage);
     }
 }
